minimal API add get cards from deck endpoind

+++ Howest.MagicCards.DAL/Models/CardDeck.cs
@@ -6,7 +6,7 @@ namespace Howest.MagicCards.DAL.Models
     public partial class CardDeck
     {
         public long CardId { get; set; }
-        public int DeckId { get; set; }
+        public long DeckId { get; set; }
         public int Amount { get; set; }
